Refactor CSSLHelperto Eliminate 64 bit warnings
Mostly replacing int with size_t.

- Updated return type of TracePacket from int to size_t.
- Changed ExtraBytes and TracedBytes from int to size_t.
- Simplified debug messages for buffer space and overflow.
- Updated unusedBytes to size_t.
- Changed extensionDataLength from UINT16 to size_t.

+++ Common/Include/SSLHelper.h
@@ -17,7 +17,7 @@ public:
 	~CSSLHelper() = default;
 	// Max length of handshake data buffer
 	void TraceHandshake();
-	static int TracePacket(const void * const Ptr, const int MaxBufBytes);
+	static size_t TracePacket(const void * const Ptr, const size_t MaxBufBytes);
 	// Is this packet a complete client initialize packet
 	bool IsClientInitialize() const;
 	// Get SNI provided hostname
